@@ -3,9 +3,10 @@ package main
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io/fs"
 	"log"
 	"os"
+	"os/user"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -13,13 +14,13 @@ import (
 	processInfo "github.com/shirou/gopsutil/process"
 )
 
-func checkAndCreateInstanceLock(path, name string) (pid int32, err error) {
-	lockFilePath := filepath.Join(dataRoot.Path, path, fmt.Sprintf("%s-lock.pid", name))
+func checkAndCreateInstanceLock(path, name string, perUser bool) (pid int32, err error) {
+	lockFilePath := getLockFilePath(path, name, perUser)
 
 	// read current pid file
-	data, err := ioutil.ReadFile(lockFilePath)
+	data, err := os.ReadFile(lockFilePath)
 	if err != nil {
-		if os.IsNotExist(err) {
+		if errors.Is(err, fs.ErrNotExist) {
 			// create new lock
 			return 0, createInstanceLock(lockFilePath)
 		}
@@ -78,7 +79,7 @@ func createInstanceLock(lockFilePath string) error {
 
 	// create lock file
 	// TODO: Investigate required permissions.
-	err = ioutil.WriteFile(lockFilePath, []byte(fmt.Sprintf("%d", os.Getpid())), 0o0666) //nolint:gosec
+	err = os.WriteFile(lockFilePath, []byte(strconv.Itoa(os.Getpid())), 0o0666) //nolint:gosec
 	if err != nil {
 		return err
 	}
@@ -86,7 +87,23 @@ func createInstanceLock(lockFilePath string) error {
 	return nil
 }
 
-func deleteInstanceLock(path, name string) error {
-	lockFilePath := filepath.Join(dataRoot.Path, path, fmt.Sprintf("%s-lock.pid", name))
-	return os.Remove(lockFilePath)
+func deleteInstanceLock(path, name string, perUser bool) error {
+	return os.Remove(getLockFilePath(path, name, perUser))
+}
+
+func getLockFilePath(path, name string, perUser bool) string {
+	if !perUser {
+		return filepath.Join(dataRoot.Path, path, fmt.Sprintf("%s-lock.pid", name))
+	}
+
+	// Get user ID for per-user lock file.
+	var userID string
+	usr, err := user.Current()
+	if err != nil {
+		log.Printf("failed to get current user: %s\n", err)
+		userID = "no-user"
+	} else {
+		userID = usr.Uid
+	}
+	return filepath.Join(dataRoot.Path, path, fmt.Sprintf("%s-%s-lock.pid", name, userID))
 }
